Feature/#52 improve pagination

See merge request 2021ss-sepm-pr-group/21ss-sepm-pr-qse-08!17

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/ProductService.java
@@ -61,8 +61,9 @@ public interface ProductService {
      * Retrieve the number of all added products.
      *
      * @return retrieves the number of all added products
+     * @throws RuntimeException occurs during database operations
      */
-    int getProductsCount();
+    Long getProductsCount();
 
     /**
      * Deletes a specific product with the given productId.
